requirements.txt
pandas
Jinja2
typer
pywin32; platform_system == "Windows"
openpyxl
beautifulsoup4
pytz
Flask
sendgrid
python-dotenv

